refactor: move `AdsDataTypeArrayInfo` and `AdsAttribute` into their own modules
