+++ packages/tcads-core/src/ads/symbol/mod.rs
@@ -1,13 +1,14 @@
+pub mod attributes;
+pub mod data_type_array_info;
 pub mod data_type_flags;
 pub mod data_type_id;
 pub mod data_type_info;
-pub mod sub_types;
 pub mod upload_info;
 
+use super::error;
+pub use attributes::AdsAttribute;
+pub use data_type_array_info::AdsDataTypeArrayInfo;
 pub use data_type_flags::AdsDataTypeFlags;
 pub use data_type_id::AdsDataTypeId;
 pub use data_type_info::AdsDataTypeInfo;
-pub use sub_types::{AdsAttribute, AdsDataTypeArrayInfo};
 pub use upload_info::AdsSymbolUploadInfo;
-
-use super::error;
